+++ NEWS.md
@@ -1,6 +1,7 @@
 # dbx 0.2.1 [unreleased]
 
 - Added `sslmode` and `sslrootcert` for RPostgreSQL
+- Fixed error with dplyr check
 
 # dbx 0.2.0